requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
constantly==23.10.4
cryptography==45.0.3
daphne==4.2.0
Django==5.2.1
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
flake8==7.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
Markdown==3.8
mccabe==0.7.0
msgpack==1.1.0
phonenumbers==9.0.6
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
pyOpenSSL==25.1.0
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
rest-framework-simplejwt==0.0.2
rpds-py==0.25.1
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
uritemplate==4.1.1
zope.interface==7.2
